Fix magento 2 auto-install provision script
This directory seems to (always?) already exist as a symlink, so check this case before trying to symlink it ourselves

+++ vagrant/provisioning/magento2-install.sh
@@ -19,7 +19,7 @@ $AS_USER chmod 755 bin/magento
 echo "Installing ..."
 
 [ -f ${HOME_DIR}/${MAGENTO_DIR}/app/etc/env.php ] || $AS_USER bin/magento setup:install --db-host=localhost --db-name=magento2 --db-user=app --db-password=${MYSQLPASSWORD} --admin-firstname=Admin --admin-lastname=user --admin-user=admin --admin-password=Password123 --admin-email=ivan@ecomdev.org --base-url=http://${VAGRANT_HOSTNAME}/ --language=en_US --timezone=Europe/Amsterdam --currency=EUR --use-rewrites=1
-$AS_USER ln -fs ../${MAGENTO_DIR}/pub/* ../public
+[ -L ../public ] || $AS_USER ln -fs ../${MAGENTO_DIR}/pub/* ../public
 $AS_USER bin/magento setup:static-content:deploy
 
 echo "Static content deployed ..."
